Change Registrar to API
