+++ apps/api/lib/api/user/user.ex
@@ -1,0 +1,15 @@
+defmodule API.User do
+  @moduledoc """
+  Context for users
+  """
+
+  alias API.User.{Authorization}
+
+  def create_user(attrs) do
+    attrs
+    |> Authorization.hash_password
+    |> DB.User.create_user
+  end
+
+  def login(attrs), do: Authorization.authorize(attrs)
+end
