+++ apps/api/priv/gettext/errors.pot
