+++ apps/api/test/api/web/controllers/user_controller_test.exs
@@ -1,0 +1,29 @@
+defmodule API.Web.UserControllerTest do
+  use API.Web.ConnCase, async: false
+
+  @email "email@email.com"
+  @password "password"
+
+  test "POST /user", %{conn: conn} do
+    res = post(conn, "/user", %{email: @email, password: @password})
+    assert res.status == 201
+  end
+
+  test "POST /user/login with correct credentials", %{conn: conn} do
+    API.User.create_user(%{"email" => @email, "password" => @password})
+    res = post(conn, "/user/login", %{email: @email, password: @password})
+    assert res.status == 200
+
+    headers = res.resp_headers
+      |> Enum.into(%{})
+
+    assert headers["authorization"]
+    assert headers["x-expires"]
+  end
+
+  test "POST /user/login with incorrect credentials", %{conn: conn} do
+    API.User.create_user(%{"email" => @email, "password" => @password})
+    res = post(conn, "/user/login", %{email: @email, password: "hacking"})
+    assert res.status == 401
+  end
+end
