--- apps/registrar/lib/registrar/user/user.ex
@@ -1,15 +1,0 @@
-defmodule Registrar.User do
-  @moduledoc """
-  Context for users
-  """
-
-  alias Registrar.User.{Authorization}
-
-  def create_user(attrs) do
-    attrs
-    |> Authorization.hash_password
-    |> DB.User.create_user
-  end
-
-  def login(attrs), do: Authorization.authorize(attrs)
-end
