--- apps/registrar/mix.exs
@@ -1,43 +1,0 @@
-defmodule Registrar.Mixfile do
-  use Mix.Project
-
-  def project do
-    [app: :registrar,
-     version: "0.0.1",
-     build_path: "../../_build",
-     config_path: "../../config/config.exs",
-     deps_path: "../../deps",
-     lockfile: "../../mix.lock",
-     elixir: "~> 1.4",
-     elixirc_paths: elixirc_paths(Mix.env),
-     compilers: [:phoenix, :gettext] ++ Mix.compilers,
-     start_permanent: Mix.env == :prod,
-     deps: deps()]
-  end
-
-  # Configuration for the OTP application.
-  #
-  # Type `mix help compile.app` for more information.
-  def application do
-    [mod: {Registrar.Application, []},
-     extra_applications: [:comeonin, :db, :logger, :guardian, :runtime_tools]]
-  end
-
-  # Specifies which paths to compile per environment.
-  defp elixirc_paths(:test), do: ["lib", "test/support"]
-  defp elixirc_paths(_),     do: ["lib"]
-
-  # Specifies your project dependencies.
-  #
-  # Type `mix help deps` for examples and options.
-  defp deps do
-    [{:phoenix, "~> 1.3.0-rc", override: true},
-     {:phoenix_pubsub, "~> 1.0"},
-     {:gettext, "~> 0.11"},
-     {:cowboy, "~> 1.0"},
-     {:comeonin, "~> 3.2"},
-     {:db, in_umbrella: true},
-     {:guardian, "~> 0.14"},
-     {:credo, "~> 0.8.4"}]
-  end
-end
